Improved the model for roles.  Added in roleName, created, and updated.
Also reduced the rights properties.  Now there is only a property per
catagory of rights.  Each category is set to the type of Mixed which
means it can be specified at runtime.

This allows us to populate each role document with only the properties
they supply to the user account; instead of a huge list of boolean
properties set to false.